# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Academia World API",
        "description": "API for Academia World - Academic Event Management Platform",
        "contact": {
            "name": "Academia World Support",
            "email": "support@academiaworld.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/v1/admin/analytics/overview": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Get platform overview statistics",
                "description": "Retrieve comprehensive platform statistics for the specified time period",
                "operationId": "getPlatformOverview",
                "parameters": [
                    {
                        "name": "days",
                        "in": "query",
                        "description": "Number of days to include in statistics",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 365,
                            "minimum": 1,
                            "example": 30
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Platform overview retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Platform overview retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "total_users": {
                                                    "type": "integer"
                                                },
                                                "total_events": {
                                                    "type": "integer"
                                                },
                                                "total_registrations": {
                                                    "type": "integer"
                                                },
                                                "active_users": {
                                                    "type": "integer"
                                                },
                                                "growth_metrics": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied - Admin privileges required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Access denied"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/analytics/user-engagement": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Get user engagement analytics",
                "description": "Retrieve user engagement metrics for the specified date range",
                "operationId": "getUserEngagementAnalytics",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Start date for analytics (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "End date for analytics (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User engagement analytics retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User engagement analytics retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "metrics": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "summary": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/analytics/event-engagement": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Get event engagement analytics",
                "description": "Retrieve event engagement metrics for the specified date range",
                "operationId": "getEventEngagementAnalytics",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Start date for analytics (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "End date for analytics (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event engagement analytics retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Event engagement analytics retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "metrics": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "summary": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/analytics/forum-activity": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Get forum activity analytics",
                "description": "Retrieve forum activity metrics for the specified date range",
                "operationId": "getForumActivityAnalytics",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Start date for analytics (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "End date for analytics (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Forum activity analytics retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forum activity analytics retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "metrics": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "summary": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}